fixup! Forward port CM Screen Security settings (1/2)

+++ packages/SystemUI/compose/features/src/com/android/systemui/bouncer/ui/composable/PatternBouncer.kt
@@ -424,7 +424,7 @@ private suspend fun showFailureAnimation(
 private fun offset(
     availableSize: Int,
     spacingPerDot: Float,
-    dotCount: Int,
+    dotCount: Byte,
     isCentered: Boolean = false,
 ): Float {
     val default = availableSize - spacingPerDot * dotCount
